Moved subscription management commands to features
