+++ commands/management.py
@@ -4,165 +4,10 @@ import logging
 import nextcord
 import pytube
 from .helper import handle_command_exception
-from bot import client, ADMINS, HOME_GUILDS
+from bot import client
 from nextcord import Interaction, Permissions, SlashOption
 
-
 _logger = logging.getLogger("main")
-
-
-@client.slash_command(
-    name="add_server",
-    description="Grant permission to a new discord server.",
-    default_member_permissions=Permissions(administrator=True),
-    guild_ids=HOME_GUILDS,
-    dm_permission=False,
-)
-async def add_server(interaction: Interaction, id: str = SlashOption(required=True)):
-    try:
-        _logger.info(
-            "commands/management: Command 'add_server' was called by "
-            + f"'{interaction.user.name}' ({interaction.user.id}) "
-            + f"in '{interaction.guild.name}' ({interaction.guild_id}) args: id:{id}"
-        )
-        if interaction.user.id not in ADMINS:
-            _logger.debug(
-                "commands/management: User does not have permission to use command 'add_server' "
-                + f"user: '{interaction.user.name}' ({interaction.user.id}) "
-                + f"guild: '{interaction.guild.name}' ({interaction.guild_id})"
-            )
-            await interaction.send(
-                "You don't have enough permissions to use this command.", ephemeral=True
-            )
-            return
-
-        interaction_response = await interaction.send("Please wait...", ephemeral=True)
-
-        server_id = int(id)
-        target_guild = client.get_guild(server_id)
-        if target_guild is None:
-            _logger.debug(f"commands/management: Server with id: {id} does not exist.")
-            await interaction_response.edit("Server with this id does not exist.")
-            return
-
-        result = data.server_add(str(target_guild), server_id)
-        if result == server_id:
-            _logger.info(
-                f"commands/management: Server '{target_guild.name}' ({target_guild.id}) "
-                + "has been registered and activated."
-            )
-            await interaction_response.edit(
-                f"Server **{target_guild}** has been registered and activated.",
-            )
-        else:
-            await interaction_response.edit(str(result))
-    except:
-        await handle_command_exception("add_server", interaction, interaction_response)
-
-
-@client.slash_command(
-    name="edit_server",
-    description="Edit permissions of a discord server.",
-    default_member_permissions=Permissions(administrator=True),
-    guild_ids=HOME_GUILDS,
-    dm_permission=False,
-)
-async def edit_server(
-    interaction: Interaction,
-    id: str = SlashOption(required=True),
-    active: bool = SlashOption(required=True),
-):
-    try:
-        _logger.info(
-            "commands/management: Command 'edit_server' was called by "
-            + f"'{interaction.user.name}' ({interaction.user.id}) "
-            + f"in '{interaction.guild.name}' ({interaction.guild_id}) args: id:{id} active:{active}"
-        )
-        if interaction.user.id not in ADMINS:
-            _logger.debug(
-                "commands/management: User does not have permission to use command 'edit_server' "
-                + f"user: '{interaction.user.name}' ({interaction.user.id}) "
-                + f"guild: '{interaction.guild.name}' ({interaction.guild_id})"
-            )
-            await interaction.send(
-                "You don't have enough permissions to use this command.", ephemeral=True
-            )
-            return
-
-        interaction_response = await interaction.send("Please wait...", ephemeral=True)
-
-        server_id = int(id)
-        target_guild = client.get_guild(server_id)
-        if target_guild is None:
-            _logger.debug(f"commands/management: Server with id: {id} does not exist.")
-            await interaction_response.edit("Server with this id does not exist.")
-            return
-
-        result = data.server_edit(server_id, active)
-        if result == server_id:
-            active_status = "activated" if active else "deactivated"
-            _logger.info(
-                f"commands/management: Server '{target_guild.name}' ({target_guild.id}) "
-                + f"has been {active_status}."
-            )
-            await interaction_response.edit(
-                f"Server **{target_guild}** has been **{active_status}**.",
-            )
-        else:
-            await interaction_response.edit(str(result))
-    except:
-        await handle_command_exception("edit_server", interaction, interaction_response)
-
-
-@client.slash_command(
-    name="remove_server",
-    description="Remove permission from a discord server.",
-    default_member_permissions=Permissions(administrator=True),
-    guild_ids=HOME_GUILDS,
-    dm_permission=False,
-)
-async def remove_server(interaction: Interaction, id: str = SlashOption(required=True)):
-    try:
-        _logger.info(
-            "commands/management: Command 'remove_server' was called by "
-            + f"'{interaction.user.name}' ({interaction.user.id}) "
-            + f"in '{interaction.guild.name}' ({interaction.guild_id}) args: id:{id}"
-        )
-        if interaction.user.id not in ADMINS:
-            _logger.debug(
-                "commands/management: User does not have permission to use command 'remove_server' "
-                + f"user: '{interaction.user.name}' ({interaction.user.id}) "
-                + f"guild: '{interaction.guild.name}' ({interaction.guild_id})"
-            )
-            await interaction.send(
-                "You don't have enough permissions to use this command.", ephemeral=True
-            )
-            return
-
-        interaction_response = await interaction.send("Please wait...", ephemeral=True)
-
-        server_id = int(id)
-        target_guild = client.get_guild(server_id)
-        if target_guild is None:
-            _logger.debug(f"commands/management: Server with id: {id} does not exist.")
-            await interaction_response.edit("Server with this id does not exist.")
-            return
-
-        result = data.server_remove(server_id)
-        if result == server_id:
-            _logger.info(
-                f"commands/management: Server '{target_guild.name}' ({target_guild.id}) "
-                + "has been removed."
-            )
-            await interaction_response.edit(
-                f"Server **{target_guild}** has been removed.",
-            )
-        else:
-            await interaction_response.edit(str(result))
-    except:
-        await handle_command_exception(
-            "remove_server", interaction, interaction_response
-        )
 
 
 @client.slash_command(
@@ -569,7 +414,7 @@ async def set_role_emoji(
             + f"message_link:{message_link} emoji_name:{emoji_name} role_name:{role_name}"
         )
         interaction_response = await interaction.send(
-            f"Please wait ...", ephemeral=True
+            f"Please wait...", ephemeral=True
         )
 
         link_parts = process_discord_message_link(message_link)
@@ -706,7 +551,7 @@ async def remove_role_message(
             + f"in '{interaction.guild.name}' ({interaction.guild_id}) args: message_link:{message_link}"
         )
         interaction_response = await interaction.send(
-            f"Please wait ...", ephemeral=True
+            f"Please wait...", ephemeral=True
         )
 
         link_parts = process_discord_message_link(message_link)
