+++ features/management/__init__.py
@@ -1,2 +1,3 @@
 from . import admin
+from . import owner
 from .feature import *
