+++ features/management/owner/subscription/feature.py
@@ -1,5 +1,4 @@
 import logging
-from . import owner
 
 _logger = logging.getLogger("main")
 
@@ -16,5 +15,4 @@ def is_active():
 def activate():
     global _active
     _active = True
-    _logger.debug("features: Feature has been activated: 'management'")
-    owner.activate()
+    _logger.debug("features: Feature has been activated: 'management/owner/subscription'")
